feat(auth): refactor hooks to use centralized auth thunks and improve dispatch usage
