+++ src/application/auth/hooks/useEmailConfirm.ts
@@ -1,9 +1,8 @@
-import { useDispatch } from "react-redux";
 import { useCallback, useState } from "react";
 
-import { AppDispatch } from "@/infrastructure/redux/store";
-import { confirmEmailThunk } from "../thunks/confirmEmailThunk";
 import { AppErrorProps } from "@/shared/errors/types";
+import { confirmEmailThunk } from "../thunks/authThunks";
+import { useAppDispatch } from "@/infrastructure/redux/hooks";
 
 export type EmailConfirmStatus = "loading" | "success" | "error";
 
@@ -12,7 +11,7 @@ export const useEmailConfirm = (): {
   errorMessage: string | null;
   confirmEmail: (token: string) => Promise<void>;
 } => {
-  const dispatch = useDispatch<AppDispatch>();
+  const dispatch = useAppDispatch();
   const [status, setStatus] = useState<EmailConfirmStatus>("loading");
   const [errorMessage, setErrorMessage] = useState<string | null>(null);
