+++ src/application/auth/hooks/useSignUp.ts
@@ -1,11 +1,10 @@
 import { useRouter } from "next/navigation";
-import { useDispatch, useSelector } from "react-redux";
 import { toast } from "sonner";
 
-import { signUpThunk } from "../thunks/signUpThunk";
 import { errorTexts, successTexts } from "@/shared/locales/messages";
-import { AppDispatch, RootState } from "@/infrastructure/redux/store";
 import { ROUTES } from "@/infrastructure/config/routes";
+import { useAppDispatch, useAppSelector } from "@/infrastructure/redux/hooks";
+import { signUpThunk } from "../thunks/authThunks";
 
 interface SignUpFormProps {
   email: string;
@@ -14,8 +13,8 @@ interface SignUpFormProps {
 
 export const useSignUp = () => {
   const router = useRouter();
-  const dispatch = useDispatch<AppDispatch>();
-  const loading = useSelector((state: RootState) => state.auth.loading);
+  const dispatch = useAppDispatch();
+  const loading = useAppSelector((state) => state.auth.loading);
 
   const signUp = async ({ email, password }: SignUpFormProps) => {
     const thunk = signUpThunk({ email, password });
